medium attack is now two kicks, heavy attack is one big punch
- changed energy to use floats
- changed heavy attack control to middle mouse button
- changed medium attack control to right mouse button
- adjusted energy costs of attacks
- changed some keyboard controls

+++ Assets/Scripts/FMOD/AnimationEvents.cs
@@ -76,7 +76,7 @@ public class AnimationEvents : MonoBehaviour
     public void ReduceEnergy(int LMH) //Light = 0, Medium = 1, Heavy = 0
     {
         //_characterAttack = transform.root.transform.Find("PlayerAttack").GetComponent<CharacterAttack>();
-        int energyUsed = 0;
+        float energyUsed = 0;
         switch (LMH)
         {
             case 0:
